refactor(request): extract parent-contact-resolution into application module
